+++ lab7/include/fs/uartfs.h
@@ -10,7 +10,7 @@ extern struct file_operations uartfs_f_ops;
 extern struct mount* uartfs_mount(struct filesystem_type*, struct dentry*);
 extern struct dentry* uartfs_create(struct dentry *, const char*);
 extern struct dentry *uartfs_lookup(struct dentry *, char*);
-extern loff_t uartfs_lseek64(struct file *, loff_t, int);
+extern int uartfs_lseek64(struct file *, loff_t, int);
 extern long uartfs_read(struct file *, char *, size_t, loff_t *);
 extern long uartfs_write(struct file *, char *, size_t, loff_t *);
 extern struct file* uartfs_open(struct dentry *,uint32_t , umode_t );
